requirements.txt
# Core Framework
google-adk
PyYAML
requests
pandas
jupyterlab
ipywidgets
Pillow # For image handling
PyPDF2 # For PDF text extraction

# Development & Testing
pytest
pytest-mock

# KiCad Utilities (Needed for parsing)
kiutils

# Optional LLM/API Clients (Uncomment and install specific ones used)
# openai
# google-cloud-aiplatform
google-generativeai # For the implemented GeminiCapability

# Other utilities
pathspec # For gitignore style matching if needed
pydantic>=2.0 # Specify Pydantic v2+ for ConfigDict etc.